refactor(trusty-agents-common): extract agent compose/deploy/manifest machinery from trusty-mpm for cross-crate reuse (refs #2892)
Moves the extends:-inheritance composer (agent_builder), the checksum +
atomic-write ownership manifest (agent_manifest), the deploy pipeline
(agent_deployer), and the shared frontmatter key/value parser out of
trusty-mpm's binary crate into a new trusty-agents-common::agents module,
mirroring the precedent set by the ToolExecutor/AgentRunner extraction.
trusty-mpm re-exports every item from its existing core::agent_builder /
core::agent_manifest / core::agent_deployer / core::frontmatter module
paths for source compatibility, so every call site compiles unchanged.

agent_manifest's Result/Error type (previously trusty-mpm's core::error::Error)
is replaced with a self-contained ManifestError (thiserror) so the shared
crate carries no host-crate dependency; trusty-mpm's core::error::Error
gains a Manifest(#[from] ManifestError) variant so existing ?-propagating
call sites keep compiling. deployer::is_agent_file widens from pub(crate)
to pub so trusty-mpm's agent_reset module can keep calling it cross-crate.

This unblocks trusty-code consuming the same agent catalog machinery
instead of forking it (#2892).
